Introduce a compatible ExpressJS webapp
It works the same as the Koa app but it's an experiment to play with
express.

I didn't enjoy it as much as Koa.
I like how Koa handles the middleware better.

Express has a lot more of dependencies, I don't know if that's good or bad.
The body parsers weren't as easy to implement.
Express is more popular and more "standard" in the community.
I'm torn.